backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
pdfplumber==0.11.4
docx2txt==0.8
psycopg2-binary==2.9.9
cryptography==42.0.8
openai==1.40.0
langchain-core==0.2.29
langchain-openai==0.1.20
weasyprint==61.2
